feat(h5): 同步 normal-modules 关于 vue2 组件的判断
