+++ packages/taro-webpack5-runner/src/plugins/TaroComponentsExportsPlugin.ts
@@ -4,9 +4,10 @@ import { toDashed } from '@tarojs/shared'
 import { componentConfig } from '../utils/component'
 
 import type { Func } from '@tarojs/taro/types/compile'
+import type AcornWalk from 'acorn-walk'
 import type { Compiler, NormalModule } from 'webpack'
 
-const walk = require('acorn-walk')
+const walk = require('acorn-walk') as typeof AcornWalk
 const NullDependency = require('webpack/lib/dependencies/NullDependency')
 
 const PLUGIN_NAME = 'TaroComponentsExportsPlugin'
@@ -16,6 +17,21 @@ interface IOptions {
   onParseCreateElement?: Func
 }
 
+export function isRenderNode (node: acorn.Node, ancestors: any = []): boolean {
+  let renderFn
+  const hasRenderMethod = ancestors.some((ancestor) => {
+    if (ancestor.type === 'FunctionExpression' && ancestor?.id?.name === 'render') {
+      renderFn = ancestor.params[0]?.name
+      return true
+    } else {
+      return false
+    }
+  })
+  // @ts-ignore
+  return hasRenderMethod && node.callee.name === renderFn
+}
+
+// TODO 合并 TaroNormalModulesPlugin 逻辑
 export default class TaroComponentsExportsPlugin {
   onParseCreateElement?: Func
   #componentsExports: Set<string>
@@ -31,8 +47,10 @@ export default class TaroComponentsExportsPlugin {
       normalModuleFactory.hooks.parser.for('javascript/auto').tap(PLUGIN_NAME, (parser) => {
         parser.hooks.program.tap(PLUGIN_NAME, (program) => {
           walk.simple(program, {
-            CallExpression: node => {
+            CallExpression: (node, ancestors) => {
+              // @ts-ignore
               const callee = node.callee
+
               if (callee.type === 'MemberExpression') {
                 if (callee.property.name !== 'createElement') {
                   return
@@ -40,23 +58,24 @@ export default class TaroComponentsExportsPlugin {
               } else {
                 const nameOfCallee = callee.name
                 if (
-                  // 兼容 react17 new jsx transtrom
+                  // 兼容 react17 new jsx transform
                   nameOfCallee !== '_jsx' && nameOfCallee !== '_jsxs' &&
                   // 兼容 Vue 3.0 渲染函数及 JSX
                   !(nameOfCallee && nameOfCallee.includes('createVNode')) &&
                   !(nameOfCallee && nameOfCallee.includes('createBlock')) &&
                   !(nameOfCallee && nameOfCallee.includes('createElementVNode')) &&
                   !(nameOfCallee && nameOfCallee.includes('createElementBlock')) &&
-                  !(nameOfCallee && nameOfCallee.includes('resolveComponent')) // 收集使用解析函数的组件名称
-                  // TODO: 兼容 vue 2.0 渲染函数及 JSX，函数名 h 与 _c 在压缩后太常见，需要做更多限制后才能兼容
-                  // nameOfCallee !== 'h' && nameOfCallee !== '_c'
+                  !(nameOfCallee && nameOfCallee.includes('resolveComponent')) && // 收集使用解析函数的组件名称
+                  // 兼容 Vue 2.0 渲染函数及 JSX
+                  !isRenderNode(node, ancestors)
                 ) {
                   return
                 }
               }
 
+              // @ts-ignore
               const type = node.arguments[0]
-              if (type.value) {
+              if (type?.value) {
                 this.onParseCreateElement?.(type.value, componentConfig)
                 this.#componentsExports.add(type.value)
               }
