+++ packages/taro-webpack5-runner/src/plugins/TaroNormalModulesPlugin.ts
@@ -1,26 +1,15 @@
 import TaroSingleEntryDependency from '../dependencies/TaroSingleEntryDependency'
 import { componentConfig, componentNameSet, elementNameSet } from '../utils/component'
+import { isRenderNode } from './TaroComponentsExportsPlugin'
 import TaroNormalModule, { TaroBaseNormalModule } from './TaroNormalModule'
 
 import type { Func } from '@tarojs/taro/types/compile'
+import type AcornWalk from 'acorn-walk'
 import type { Compiler } from 'webpack'
 
-const walk = require('acorn-walk')
+const walk = require('acorn-walk') as typeof AcornWalk
 
 const PLUGIN_NAME = 'TaroNormalModulesPlugin'
-
-function isRenderNode (node, ancestors): boolean {
-  let renderFn
-  const hasRenderMethod = ancestors.some((ancestor) => {
-    if (ancestor.type === 'FunctionExpression' && ancestor?.id?.name === 'render') {
-      renderFn = ancestor.params[0]?.name
-      return true
-    } else {
-      return false
-    }
-  })
-  return hasRenderMethod && node.callee.name === renderFn
-}
 
 export default class TaroNormalModulesPlugin {
   isCache = true
@@ -70,6 +59,7 @@ export default class TaroNormalModulesPlugin {
 
           walk.ancestor(ast, {
             CallExpression: (node, ancestors) => {
+              // @ts-ignore
               const callee = node.callee
 
               if (callee.type === 'MemberExpression') {
@@ -79,7 +69,7 @@ export default class TaroNormalModulesPlugin {
               } else {
                 const nameOfCallee = callee.name
                 if (
-                  // 兼容 react17 new jsx transtrom
+                  // 兼容 react17 new jsx transform
                   nameOfCallee !== '_jsx' && nameOfCallee !== '_jsxs' &&
                   // 兼容 Vue 3.0 渲染函数及 JSX
                   !(nameOfCallee && nameOfCallee.includes('createVNode')) &&
@@ -94,6 +84,7 @@ export default class TaroNormalModulesPlugin {
                 }
               }
 
+              // @ts-ignore
               const [type, prop] = node.arguments
               const componentName = type.name
 
@@ -121,9 +112,9 @@ export default class TaroNormalModulesPlugin {
               const props = prop.properties
                 .filter(p => p.type === 'Property' && p.key.type === 'Identifier' && p.key.name !== 'children' && p.key.name !== 'id')
               const res = props.map(p => p.key.name).join('|')
-        
+
               props.forEach(p => attrs.add(p.key.name))
-          
+
               currentModule.collectProps[type.value] = res
             },
           })
